requirement_local.txt
beautifulsoup4==4.12.3
chromadb==0.4.24
huggingface-hub==0.20.3
langchain==0.1.16
langchain-text-splitters==0.0.1
pandas==2.0.3
numpy==1.25.2
regex==2023.12.25
requests==2.31.0
sentence-transformers==2.6.1
transformers==4.38.2
streamlit==1.33.0
Pillow==9.4.0